Refactor product hotspots into focused modules
