+++ server/_core/image-generation/imageServiceErrors.ts
@@ -1,0 +1,5 @@
+export class InvalidGenerationInputError extends Error {}
+export class MissingOpenAiApiKeyError extends Error {}
+export class GenerationTimeoutError extends Error {}
+export class MissingAppBaseUrlError extends Error {}
+export class MissingObjectStorageConfigError extends Error {}
