+++ server/_core/messengerState.ts
@@ -6,19 +6,19 @@ import {
 import type { ActiveExperience } from "./activeExperience";
 import type { EntryIntent } from "./entryIntent";
 import type { Lang } from "./i18n";
-import { toUserKey } from "./privacy";
 import {
   clearStateStore,
   deleteState,
-  findInMemoryState,
-  getOrCreateStoredState,
   isPromiseLike,
-  isRedisStateStoreEnabled,
-  readState,
   type MaybePromise,
-  updateStoredState,
-  writeState,
 } from "./stateStore";
+import { toUserKey } from "./privacy";
+import { getDayKey } from "./messengerStateNormalization";
+import {
+  getOrCreatePersistedState,
+  getPersistedState,
+  patchState,
+} from "./messengerStatePersistence";
 
 export type ConversationState =
   | "IDLE"
@@ -102,220 +102,6 @@ const QUICK_REPLIES_BY_STATE: Record<ConversationState, StateQuickReply[]> = {
   ],
 };
 
-type PartialState = Partial<MessengerUserState>;
-
-type StateNormalizationBase = {
-  resolvedPsid: string;
-  fallback: MessengerUserState;
-};
-
-type LegacyStateFields = {
-  stage: MessengerFlowState;
-  lastPhoto: string | null;
-  selectedStyle: string | null;
-  lastGeneratedUrl: string | null | undefined;
-};
-
-function looksLikeUserKey(value: string): boolean {
-  return /^[a-f0-9]{64}$/i.test(value);
-}
-
-function getUserKey(psid: string): string {
-  return looksLikeUserKey(psid) ? psid : toUserKey(psid);
-}
-
-function createDefaultState(psid: string, now = Date.now()): MessengerUserState {
-  return {
-    psid,
-    userKey: getUserKey(psid),
-    stage: "IDLE",
-    state: "IDLE",
-    lastEntryIntent: null,
-    activeExperience: null,
-    lastUserMessageAt: undefined,
-    lastPhotoUrl: null,
-    lastPhoto: null,
-    lastPhotoSource: null,
-    selectedStyle: null,
-    chosenStyle: null,
-    selectedStyleCategory: null,
-    preselectedStyle: null,
-    preferredLang: "nl",
-    consentGiven: false,
-    consentTimestamp: undefined,
-    pendingDeleteConfirm: false,
-    hasSeenIntro: false,
-    pendingImageUrl: undefined,
-    pendingImageAt: undefined,
-    faceMemoryConsent: null,
-    lastSourceImageUrl: null,
-    lastSourceImageUpdatedAt: null,
-    pendingSourceImageDeleteUrl: null,
-    lastImageUrl: undefined,
-    lastGeneratedUrl: null,
-    lastStyle: undefined,
-    lastPrompt: undefined,
-    lastGeneratedAt: undefined,
-    lastVariantCursor: undefined,
-    quota: {
-      dayKey: getDayKey(now),
-      count: 0,
-    },
-    updatedAt: now,
-  };
-}
-
-function normalizeState(
-  psid: string,
-  value: PartialState | null | undefined
-): MessengerUserState {
-  const base = createStateNormalizationBase(psid, value);
-  const legacyFields = resolveLegacyStateFields(value, base.fallback);
-
-  return applyNormalizedStateShape(value, base, legacyFields);
-}
-
-function createStateNormalizationBase(
-  psid: string,
-  value: PartialState | null | undefined
-): StateNormalizationBase {
-  const resolvedPsid = value?.psid ?? psid;
-  const fallback = createDefaultState(resolvedPsid);
-  return { resolvedPsid, fallback };
-}
-
-function resolveLegacyStateFields(
-  value: PartialState | null | undefined,
-  fallback: MessengerUserState
-): LegacyStateFields {
-  const stage = value?.stage ?? value?.state ?? fallback.stage;
-  const lastPhoto =
-    value?.lastPhotoUrl ?? value?.lastPhoto ?? fallback.lastPhoto;
-  const selectedStyle =
-    value?.selectedStyle ?? value?.chosenStyle ?? fallback.selectedStyle;
-  const lastGeneratedUrl =
-    value?.lastGeneratedUrl ?? value?.lastImageUrl ?? fallback.lastGeneratedUrl;
-
-  return { stage, lastPhoto, selectedStyle, lastGeneratedUrl };
-}
-
-function applyNormalizedStateShape(
-  value: PartialState | null | undefined,
-  base: StateNormalizationBase,
-  legacyFields: LegacyStateFields
-): MessengerUserState {
-  const { resolvedPsid, fallback } = base;
-  const { stage, lastPhoto, selectedStyle, lastGeneratedUrl } = legacyFields;
-
-  return {
-    ...fallback,
-    ...value,
-    psid: resolvedPsid,
-    userKey: value?.userKey ?? fallback.userKey,
-    consentGiven: value?.consentGiven ?? fallback.consentGiven,
-    consentTimestamp: value?.consentTimestamp ?? fallback.consentTimestamp,
-    pendingDeleteConfirm:
-      value?.pendingDeleteConfirm ?? fallback.pendingDeleteConfirm,
-    hasSeenIntro: value?.hasSeenIntro ?? fallback.hasSeenIntro,
-    stage,
-    state: stage,
-    lastEntryIntent: value?.lastEntryIntent ?? fallback.lastEntryIntent,
-    activeExperience: value?.activeExperience ?? fallback.activeExperience,
-    lastUserMessageAt: value?.lastUserMessageAt ?? fallback.lastUserMessageAt,
-    lastPhotoUrl: lastPhoto,
-    lastPhoto,
-    lastPhotoSource: value?.lastPhotoSource ?? fallback.lastPhotoSource,
-    selectedStyle,
-    chosenStyle: selectedStyle,
-    selectedStyleCategory:
-      value?.selectedStyleCategory ?? fallback.selectedStyleCategory,
-    lastImageUrl: value?.lastImageUrl ?? lastGeneratedUrl ?? fallback.lastImageUrl,
-    lastGeneratedUrl,
-    faceMemoryConsent: value?.faceMemoryConsent ?? fallback.faceMemoryConsent,
-    lastSourceImageUrl: value?.lastSourceImageUrl ?? fallback.lastSourceImageUrl,
-    lastSourceImageUpdatedAt:
-      value?.lastSourceImageUpdatedAt ?? fallback.lastSourceImageUpdatedAt,
-    pendingSourceImageDeleteUrl:
-      value?.pendingSourceImageDeleteUrl ?? fallback.pendingSourceImageDeleteUrl,
-    quota: {
-      dayKey: value?.quota?.dayKey ?? fallback.quota.dayKey,
-      count: value?.quota?.count ?? fallback.quota.count,
-    },
-    updatedAt: value?.updatedAt ?? fallback.updatedAt,
-  };
-}
-
-function saveState(psid: string, nextState: MessengerUserState): MaybePromise<MessengerUserState> {
-  const result = writeState(psid, nextState);
-  if (isPromiseLike(result)) {
-    return result.then(() => nextState);
-  }
-
-  return nextState;
-}
-
-function getStateFromMemory(psid: string): MessengerUserState | null {
-  const direct = readState<PartialState>(psid);
-  if (isPromiseLike(direct)) {
-    throw new Error("Unexpected async state read in memory mode");
-  }
-
-  if (direct) {
-    return normalizeState(psid, direct);
-  }
-
-  const userKey = getUserKey(psid);
-  const legacyState = findInMemoryState<PartialState>(state => state.userKey === userKey);
-  return legacyState ? normalizeState(legacyState.psid ?? psid, legacyState) : null;
-}
-
-function getStateFromRedis(psid: string): Promise<MessengerUserState | null> {
-  return Promise.resolve(readState<PartialState>(psid)).then(state => {
-    return state ? normalizeState(psid, state) : null;
-  });
-}
-
-function patchStateInMemory(psid: string, patch: PartialState, now = Date.now()): MessengerUserState {
-  const current = getOrCreateState(psid);
-  if (isPromiseLike(current)) {
-    throw new Error("Unexpected async state patch in memory mode");
-  }
-
-  const nextState = normalizeState(psid, {
-    ...current,
-    ...patch,
-    updatedAt: now,
-  });
-
-  const saved = saveState(psid, nextState);
-  if (isPromiseLike(saved)) {
-    throw new Error("Unexpected async state save in memory mode");
-  }
-
-  return saved;
-}
-
-function patchStateInRedis(psid: string, patch: PartialState, now = Date.now()): Promise<MessengerUserState> {
-  return Promise.resolve(
-    updateStoredState<PartialState>(psid, current => {
-      const nextState = normalizeState(psid, {
-        ...normalizeState(psid, current),
-        ...patch,
-        updatedAt: now,
-      });
-      return nextState;
-    }),
-  ).then(state => normalizeState(psid, state));
-}
-
-function patchState(psid: string, patch: PartialState, now = Date.now()): MaybePromise<MessengerUserState> {
-  if (!isRedisStateStoreEnabled()) {
-    return patchStateInMemory(psid, patch, now);
-  }
-
-  return patchStateInRedis(psid, patch, now);
-}
-
 export function getQuickRepliesForState(state: ConversationState): StateQuickReply[] {
   return QUICK_REPLIES_BY_STATE[state];
 }
@@ -324,9 +110,7 @@ export function anonymizePsid(psid: string): string {
   return toUserKey(psid);
 }
 
-export function getDayKey(now = Date.now()): string {
-  return new Date(now).toISOString().slice(0, 10);
-}
+export { getDayKey };
 
 function getMessengerResponseWindowMs(): number {
   const configured = Number(process.env.MESSENGER_RESPONSE_WINDOW_MS);
@@ -338,11 +122,7 @@ function getMessengerResponseWindowMs(): number {
 }
 
 export function getState(psid: string): MaybePromise<MessengerUserState | null> {
-  if (!isRedisStateStoreEnabled()) {
-    return getStateFromMemory(psid);
-  }
-
-  return getStateFromRedis(psid);
+  return getPersistedState(psid);
 }
 
 export function clearUserState(psid: string): MaybePromise<void> {
@@ -370,19 +150,7 @@ export function hasOpenMessengerResponseWindow(psid: string, now = Date.now()): 
 }
 
 export function getOrCreateState(psid: string): MaybePromise<MessengerUserState> {
-  if (!isRedisStateStoreEnabled()) {
-    const state = getStateFromMemory(psid);
-    if (state) {
-      return state;
-    }
-
-    const createdState = createDefaultState(psid);
-    return saveState(psid, createdState);
-  }
-
-  return Promise.resolve(getOrCreateStoredState(psid, () => createDefaultState(psid))).then(state => {
-    return normalizeState(psid, state);
-  });
+  return getOrCreatePersistedState(psid);
 }
 
 export function setFlowState(psid: string, nextState: MessengerFlowState): MaybePromise<void> {
